[MOD] planner dto

+++ src/main/java/com/its/mergeplan_v1/controller/PlannerController.java
@@ -25,7 +25,7 @@ public class PlannerController {
     }
 
     @PatchMapping("/auth/planner/item/{plan_id}")  // 플래너 일정 수정하기
-    public Planner updatePlan(Authentication authentication, @PathVariable int plan_id, @RequestBody PostPlanner postPlanner){
+    public Planner updatePlan(Authentication authentication, @PathVariable Long plan_id, @RequestBody PostPlanner postPlanner){
         PrincipalDetails principalDetails = (PrincipalDetails) authentication.getPrincipal();
         return plannerService.updatePlan(principalDetails.getUser(), postPlanner, plan_id);
     }
@@ -38,7 +38,7 @@ public class PlannerController {
     }
 
     @DeleteMapping("/auth/planner/item/{plan_id}")  // 플랜 삭제하기
-    public void deletePlan(Authentication authentication, @PathVariable int plan_id){
+    public void deletePlan(Authentication authentication, @PathVariable Long plan_id){
         PrincipalDetails principalDetails = (PrincipalDetails) authentication.getPrincipal();
         plannerService.deletePlan(principalDetails.getUser(), plan_id);
     }
